Allow withdrawals that are larger than bucket limit
This commit adds an `allowLargerWithdrawals` option to the `RollingLimiterOptions`
which, when enabled, allows withdrawals that are larger than the bucket
limit/size IFF the bucket is full. When a withdrawal that's larger than
the bucket size is made, the remaining token amount will become negative
and future requests will need to wait until the bucket has refilled to
an appropriate amount.